Exec: rails db:seed

+++ db/migrate/20210831075201_create_record.rb
@@ -3,10 +3,10 @@ class CreateRecord < ActiveRecord::Migration[6.0]
     create_table :records do |t|
       t.references :exercise, null: false, foreign_key: true
       t.integer :name_id, null: false
-      t.integer :weight, null:false, default: 0
+      t.float :weight, null:false, default: 0
       t.integer :reps_count, null: false, default: 0
       t.text :note, default: ""
-      t.integer :RM, null: false, default: 0
+      t.float :RM, null: false, default: 0
       t.date :date, null: false
       
       t.timestamps
